Ex6 done

See merge request intern2021/course21.01/internal-course!24

+++ src/main/java/com/inetum/training/todo/controller/TodoRestController.java
@@ -39,16 +39,20 @@ public class TodoRestController {
 
     @RequestMapping(value = "/{id}", method = RequestMethod.PUT)
     public void update(@RequestBody @Valid Todo todo, @PathVariable("id") Long id){
-        todo.setId(id);
-        todoRepository.save(todo);
+        if(todoRepository.existsById(id)){
+            todo.setId(id);
+            todoRepository.save(todo);
+        }else{
+            throw new EntityNotFoundException();
+        }
     }
 
     @RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void delete(@PathVariable("id") Long id){
-        try {
+        if(todoRepository.existsById(id)){
             todoRepository.deleteById(id);
-        }catch(EmptyResultDataAccessException ex){
+        }else{
             throw new EntityNotFoundException();
         }
     }
@@ -59,9 +63,4 @@ public class TodoRestController {
         return "nie znaleziono elementu o podanym id";
     }
 
-//    private List<Todo> iterable2list(Iterable<Todo> todos) {
-//        List<Todo> result = new ArrayList<>();
-//        todos.forEach( t -> result.add(t));
-//        return result;
-//    }
 }
